// src/employees/dto/database-colabs.dto.ts
export class DatabaseColabsDto {  
  ItemInternalId: string;
  Empresa: number;
  Nome_Empresa: string;
  Tipo_Colaborador: number;
  desc_tipcol: string;
  Matricula: number;
  Nome_Colaborador: string;
  Apelido_Colaborador: string;
  Filial: number;
  Nome_Filial: string;
  Razao_Social_Filial: string;
  Centro_de_Custo: string;
  Nome_Centro_de_Custo: string;
  Cargo: string;
  Titulo_Cargo: string;
  Data_Admissao: string; 
  Data_Nascimento: string;
  Foto_Empregado: string;
}